# improve performance for number to string conversion functions