This is kinda silly but it is what it is...

Benchmark:

```jl
for base in (2, 8, 10, 16)
    @btime begin
        u = rand(UInt32)
        string(u; base=$base)
    end
end
```

Before
```jl
  385.406 ns (2 allocations: 118 bytes)
  168.925 ns (2 allocations: 96 bytes)
  131.273 ns (2 allocations: 96 bytes)
  149.896 ns (2 allocations: 96 bytes)
```

After

```jl
  72.232 ns (2 allocations: 118 bytes)
  49.118 ns (2 allocations: 96 bytes)
  61.106 ns (2 allocations: 96 bytes)
  68.188 ns (2 allocations: 96 bytes)
```