@KristofferC 2018-08-15
The following patch also manages to speed it up:

```diff
diff --git a/base/char.jl b/base/char.jl
index 700c61db27c..d071a3988bd 100644
--- a/base/char.jl
+++ b/base/char.jl
@@ -108,10 +108,14 @@ and [`show_invalid`](@ref).
 """
 isoverlong(c::AbstractChar) = false

-function UInt32(c::Char)
+@inline function UInt32(c::Char)
     # TODO: use optimized inline LLVM
     u = reinterpret(UInt32, c)
     u < 0x80000000 && return u >> 24
+    return _uint32(u)
+end
+
+@noinline function _uint32(u::UInt32)
     l1 = leading_ones(u)
     t0 = trailing_zeros(u) & 56
     (l1 == 1) | (8l1 + t0 > 32) |
@@ -142,8 +146,12 @@ that support overlong encodings should implement `Base.decode_overlong`.
 """
 decode_overlong

-function Char(u::UInt32)
+@inline function Char(u::UInt32)
     u < 0x80 && return reinterpret(Char, u << 24)
+    return _char(u)
+end
+
+@noinline function _char(u::UInt32)
     u < 0x00200000 || code_point_err(u)::Union{}
     c = ((u << 0) & 0x0000003f) | ((u << 2) & 0x00003f00) |
         ((u << 4) & 0x003f0000) | ((u << 6) & 0x3f000000)
```

It is a bit more general but I guess might slow things down if the old `Char` function inlined and you are dealing with non ASCII a lot?