@johnfgibson 2018-08-15
That is great. Any chance there's a similar fix for a performance regression in the print_to_file microbenchmark? It went from 6.86 in julia-0.6.4 to 10.8 in julia-0.7.0 and 1.0.0.  The core of the benchmark is 

```
for i = 1:n
    @printf(io, "%d %d\n", i, i + 1)
end
```