@jebej 2018-08-15
Maybe use `Int('0')` for clarity?