@KristofferC 2018-08-15
> That is great. Any chance there's a similar fix for a performance regression in the print_to_file microbenchmark? It went from 6.86 in julia-0.6.4 to 10.8 in julia-0.7.0 and 1.0.0. The core of the benchmark is

There are, in general, `String` regressions a bit here and there. I'll see if there is some simple fix that can be applied like here.