@KristofferC 2018-08-15
For now, I think we should go with this "local" performance improvement and then we can look at a perhaps more general but more intrusive change like https://github.com/JuliaLang/julia/pull/28661#issuecomment-413124171.