@jebej 2018-08-15
> That kills performance.

Argh, I've been using this type of pattern. Isn't the conversion done at compile time? Why would it be worse than just '0'?