@KristofferC 2018-08-15
> Argh, I've been using this type of pattern. Isn't the conversion done at compile time? Why would it be worse than just '0'?

It should be and in some cases it does. 

```
julia> f() = UInt8('0')
f (generic function with 1 method)

julia> @code_llvm f()

; Function f
; Location: REPL[4]:1
define i8 @julia_f_34955() {
top:
  ret i8 48
}
```

But for some reason, it doesn't seem to happen here. Not sure why.