@vtjnash 2018-08-15
Typically, we handle that case by writing it thusly:
```julia
@eval function f() = begin
    some_computation($(some_constant_computation(some_constants)))
end
```